FIN, ACK, SYN, Connect(), Accept(), Done
Now need Flow Control and Reliable Transport

+++ proj1N/src/ports.h
@@ -3,7 +3,7 @@
 
 typedef struct port{
 	bool isUsed;
-	uint8_t scktID;
+	uint16_t scktID;
 }port;
